modify type of speed variable

+++ lib/Driver/Motor/Motor.h
@@ -14,10 +14,10 @@ class Motor
 
   public:
     Motor();
-    void TurnLeft(int speed);
-    void TurnRight(int speed);
-    void Movefront(int speed);
-    void MoveBack(int speed);
+    void TurnLeft(float speed);
+    void TurnRight(float speed);
+    void Movefront(float speed);
+    void MoveBack(float speed);
 
 };
